requirements.txt
streamlit>=1.28.0
openai>=1.3.0
google-generativeai>=0.3.0
pinecone>=3.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
typing-extensions>=4.8.0